@@ -32,8 +32,10 @@ class ListCommand extends Command
         $masters = $masters->all();
 
         if (empty($masters)) {
-            return $this->info('No machines are running.');
+            return $this->components->info('No machines are running.');
         }
+
+        $this->output->writeln('');
 
         $this->table([
             'Name', 'PID', 'Supervisors', 'Status',
@@ -47,5 +49,7 @@ class ListCommand extends Command
                 $master->status,
             ];
         })->all());
+
+        $this->output->writeln('');
     }
 }
